+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `with_capacity_bytes` to `WriteBatch` (0xdeafbeef)
 * Add `set_compaction_pri` to `Options` (0xdeafbeef)
 * Add `lto` feature to enable link-time optimization using `linker-plugin-lto` (0xdeafbeef)
+* Allow to specify ttl per column family (0xdeafbeef)
 
 ## [Breaking Changes]
 * Update jemalloc-sys to 0.6.0 (0xdeafbeef)